+++ src/components/ColorPicker/ColorPicker.style.tsx
@@ -41,7 +41,7 @@ const StlyedColorPicker = styled.div`
     position: absolute;
     right: 0;
     top: 40px;
-    z-index: 2;
+    z-index: 1;
     border-radius: 5px;
     overflow: hidden;
     box-shadow: rgba(0, 0, 0, 0.15) 0px 0px 2px, rgba(0, 0, 0, 0.15) 0px 4px 8px;