+++ src/components/RadioOption/RadioOption.style.tsx
@@ -1,5 +1,5 @@
 import { styled } from '@storybook/theming';
-import { getPrimaryColor } from '../../helper';
+import { getPrimaryColor, getInverseTextColor } from '../../helper';
 
 const StyledRadioOption = styled.label`
   position: relative;
@@ -21,13 +21,13 @@ const StyledRadioOption = styled.label`
   div {
     display: flex;
     align-items: center;
-    padding: 5px;
+    padding: 5px 10px;
     margin: 0 0 0 10px;
-    border-radius: 5px;
+    border-radius: 3rem;
+    opacity: 0.5;
     color: ${({ theme }) => getPrimaryColor(theme)};
     border: 1px solid ${({ theme }) => getPrimaryColor(theme)};
-    opacity: 0.5;
-    pointer-events: none;
+    transition: all 0.2s ease-in-out;
 
     svg {
       width: 20px;
@@ -36,8 +36,14 @@ const StyledRadioOption = styled.label`
     }
   }
 
+  div:hover {
+    opacity: 1;
+  }
+
   input:checked + div {
     opacity: 1;
+    background: ${({ theme }) => getPrimaryColor(theme)};
+    color: ${({ theme }) => getInverseTextColor(theme)};
   }
 `;